I removed excess logging from portions of the engine and fixed a bug in the Util.deleteDirectory method and added a comprehensive unit test for the Util.deleteDirectory method.

+++ core/src/main/java/org/eclipse/jemo/sys/internal/Util.java
@@ -249,14 +249,21 @@ public class Util {
     }
 
     public static final boolean deleteDirectory(File dir) {
-        Arrays.asList(dir.listFiles()).stream().forEach(f -> {
-            if (f.isDirectory()) {
-                deleteDirectory(f);
-            } else {
-                f.delete();
-            }
-        });
-        return dir.delete();
+				if(dir != null) {
+					File[] fileList = dir.listFiles();
+					if(fileList != null && fileList.length > 0) {
+						Arrays.asList(fileList).stream().forEach(f -> {
+								if (f.isDirectory()) {
+										deleteDirectory(f);
+								} else {
+										f.delete();
+								}
+						});
+					}
+					return dir.delete();
+				}
+				
+				return true;
     }
 
     public static Set<Integer> parseIntegerRangeDefinition(final String rangeDefinition) {
